chore: consolidate redundant tests; add arc3 property validation tests
